feat(runtime): shared transaction executor + minimal FlightSQL entry
Extract the gated-transaction orchestrator out of the HTTP module so the HTTP
`/v1/sql` and FlightSQL entry points drive the same code, and add a FlightSQL
transaction entry (part of #11830's v1 scope: FlightSQL support).

- New `datafusion::query::transaction` module: `run_transaction` runs a
  `BEGIN … COMMIT` body atomically and returns a transport-neutral
  `TransactionOutcome` (the final statement's batches) or `TransactionError`
  (`Rejected` / `Plan` / `Query` / `Stream` / `Conflict { table }` /
  `Publish`). The prepare/execute/commit logic — participant registration,
  per-key OCC commit, post-commit cache invalidation — moves here from
  `http/v1`; each transport maps the neutral error to its own status.
- HTTP `sql_to_http_response` calls `run_transaction` and shapes the outcome
  into a response (`Conflict` → 409); response-shaping stays in `http/v1`.
- FlightSQL: `statement_query` and `prepared_statement_query` (the path the
  ADBC driver uses) accept a whole `BEGIN … COMMIT` body as one
  `CommandStatementQuery`. `do_get` runs it through `run_transaction` and
  streams the final statement's result; `get_flight_info` and
  create-prepared-statement advertise the FINAL statement's schema
  (`schema_statement`), planned without executing — so any write runs only at
  `do_get`. `Conflict` → gRPC `Aborted`; read-only principals are rejected for
  write transactions.
- Only the final statement's batches are materialized: its result can be
  emitted to the client only once the commit is confirmed (a conflict must
  surface as an error, never as a truncated result). Intermediate statements
  (the gate, earlier writes) are drained without buffering.

The full FlightSQL transaction-action protocol (a `transaction_id` correlated
across round-trips) is out of scope; the single-body form matches the HTTP
entry.

Validated end-to-end against a live runtime with two accelerated
Postgres-backed tables. HTTP: atomic two-table transfer with conservation,
cross-table atomic rollback, insufficient-funds gate, 10-way same-key
conservation, disjoint-table concurrency. FlightSQL (ADBC driver): gated
transfer, failing-gate atomic rollback, plain SELECT regression.
